// apps/cms/src/generated/contentOptions.ts
export const generatedBlogCategoryOptions = [
  {
    "label": "Blood Tests",
    "labels": {
      "en": "Blood Tests",
      "ja": "血液検査"
    },
    "value": "blood-tests"
  },
  {
    "label": "IV Therapy",
    "labels": {
      "en": "IV Therapy",
      "ja": "点滴療法"
    },
    "value": "iv-therapy"
  },
  {
    "label": "Medication",
    "labels": {
      "en": "Medication",
      "ja": "薬"
    },
    "value": "medication"
  },
  {
    "label": "Stem Cell Therapy",
    "labels": {
      "en": "Stem Cell Therapy",
      "ja": "幹細胞治療"
    },
    "value": "stem-cell-therapy"
  }
] as const

export const generatedBlogTagOptions = [
  {
    "label": "allergies",
    "value": "allergies"
  },
  {
    "label": "athletes",
    "value": "athletes"
  },
  {
    "label": "beauty",
    "value": "beauty"
  },
  {
    "label": "booking",
    "value": "booking"
  },
  {
    "label": "clarity",
    "value": "clarity"
  },
  {
    "label": "cold and flu",
    "value": "cold and flu"
  },
  {
    "label": "competition",
    "value": "competition"
  },
  {
    "label": "conditioned media",
    "value": "conditioned media"
  },
  {
    "label": "cost",
    "value": "cost"
  },
  {
    "label": "dehydration",
    "value": "dehydration"
  },
  {
    "label": "detox",
    "value": "detox"
  },
  {
    "label": "exosome iv",
    "value": "exosome iv"
  },
  {
    "label": "fatigue",
    "value": "fatigue"
  },
  {
    "label": "focus",
    "value": "focus"
  },
  {
    "label": "glutathione",
    "value": "glutathione"
  },
  {
    "label": "gut health",
    "value": "gut health"
  },
  {
    "label": "hangover iv",
    "value": "hangover iv"
  },
  {
    "label": "hydration",
    "value": "hydration"
  },
  {
    "label": "immune health",
    "value": "immune health"
  },
  {
    "label": "iv therapy",
    "value": "iv therapy"
  },
  {
    "label": "jet lag",
    "value": "jet lag"
  },
  {
    "label": "low energy",
    "value": "low energy"
  },
  {
    "label": "mobile iv therapy",
    "value": "mobile iv therapy"
  },
  {
    "label": "nausea",
    "value": "nausea"
  },
  {
    "label": "nutrition",
    "value": "nutrition"
  },
  {
    "label": "pricing",
    "value": "pricing"
  },
  {
    "label": "private care",
    "value": "private care"
  },
  {
    "label": "recovery",
    "value": "recovery"
  },
  {
    "label": "regenerative medicine",
    "value": "regenerative medicine"
  },
  {
    "label": "relaxation",
    "value": "relaxation"
  },
  {
    "label": "risks",
    "value": "risks"
  },
  {
    "label": "safety",
    "value": "safety"
  },
  {
    "label": "skin iv",
    "value": "skin iv"
  },
  {
    "label": "stem cell therapy",
    "value": "stem cell therapy"
  },
  {
    "label": "stress",
    "value": "stress"
  },
  {
    "label": "tokyo",
    "value": "tokyo"
  },
  {
    "label": "training recovery",
    "value": "training recovery"
  },
  {
    "label": "travel",
    "value": "travel"
  },
  {
    "label": "vitamins",
    "value": "vitamins"
  },
  {
    "label": "weight management",
    "value": "weight management"
  },
  {
    "label": "wellness",
    "value": "wellness"
  },
  {
    "label": "アレルギー",
    "value": "アレルギー"
  },
  {
    "label": "ウェルネス",
    "value": "ウェルネス"
  },
  {
    "label": "エクソソーム IV",
    "value": "エクソソーム IV"
  },
  {
    "label": "グルタチオン",
    "value": "グルタチオン"
  },
  {
    "label": "スキン IV",
    "value": "スキン IV"
  },
  {
    "label": "ストレス",
    "value": "ストレス"
  },
  {
    "label": "デトックス",
    "value": "デトックス"
  },
  {
    "label": "トレーニング後の回復",
    "value": "トレーニング後の回復"
  },
  {
    "label": "ビタミン",
    "value": "ビタミン"
  },
  {
    "label": "リスク",
    "value": "リスク"
  },
  {
    "label": "リラクゼーション",
    "value": "リラクゼーション"
  },
  {
    "label": "予約",
    "value": "予約"
  },
  {
    "label": "二日酔い IV",
    "value": "二日酔い IV"
  },
  {
    "label": "低エネルギー",
    "value": "低エネルギー"
  },
  {
    "label": "体重管理",
    "value": "体重管理"
  },
  {
    "label": "価格設定",
    "value": "価格設定"
  },
  {
    "label": "免疫の健康",
    "value": "免疫の健康"
  },
  {
    "label": "再生医療",
    "value": "再生医療"
  },
  {
    "label": "吐き気",
    "value": "吐き気"
  },
  {
    "label": "回復",
    "value": "回復"
  },
  {
    "label": "在宅介護",
    "value": "在宅介護"
  },
  {
    "label": "安全",
    "value": "安全"
  },
  {
    "label": "幹細胞治療",
    "value": "幹細胞治療"
  },
  {
    "label": "旅行",
    "value": "旅行"
  },
  {
    "label": "明瞭さ",
    "value": "明瞭さ"
  },
  {
    "label": "時差ぼけ",
    "value": "時差ぼけ"
  },
  {
    "label": "東京",
    "value": "東京"
  },
  {
    "label": "栄養",
    "value": "栄養"
  },
  {
    "label": "水分補給",
    "value": "水分補給"
  },
  {
    "label": "焦点",
    "value": "焦点"
  },
  {
    "label": "疲労",
    "value": "疲労"
  },
  {
    "label": "移動式点滴療法",
    "value": "移動式点滴療法"
  },
  {
    "label": "競技会",
    "value": "競技会"
  },
  {
    "label": "美",
    "value": "美"
  },
  {
    "label": "脱水",
    "value": "脱水"
  },
  {
    "label": "腸の健康",
    "value": "腸の健康"
  },
  {
    "label": "調製培地",
    "value": "調製培地"
  },
  {
    "label": "費用",
    "value": "費用"
  },
  {
    "label": "選手",
    "value": "選手"
  },
  {
    "label": "静脈内療法",
    "value": "静脈内療法"
  },
  {
    "label": "風邪とインフルエンザ",
    "value": "風邪とインフルエンザ"
  }
] as const

export const generatedRelatedServiceSlugOptions = [
  {
    "label": "AGA Medication",
    "value": "androgenetic-alopecia-medicine"
  },
  {
    "label": "Blood Tests",
    "value": "blood-tests"
  },
  {
    "label": "ED Medication",
    "value": "ed-medication"
  },
  {
    "label": "Energy & Fatigue Recovery IV",
    "value": "energy-fatigue-recovery-iv"
  },
  {
    "label": "Exosome IV Drip",
    "value": "exosome-iv-drip"
  },
  {
    "label": "Hangover IV Drip",
    "value": "hangover-iv-drip"
  },
  {
    "label": "Hormone Blood Testing",
    "value": "hormone-blood-testing"
  },
  {
    "label": "Immune Boost IV Therapy",
    "value": "immune-boost-iv-therapy"
  },
  {
    "label": "IV Therapy",
    "value": "iv-therapy"
  },
  {
    "label": "IV Vitamin Therapy",
    "value": "iv-vitamin-therapy"
  },
  {
    "label": "Medication",
    "value": "medication"
  },
  {
    "label": "Nutrition Blood Testing",
    "value": "nutrition-blood-testing"
  },
  {
    "label": "Skin Brightening IV Drip",
    "value": "skin-brightening-iv-drip"
  },
  {
    "label": "Stem Cell Nasal Spray",
    "value": "stem-cell-nasal-spray"
  },
  {
    "label": "Stem Cell Therapy",
    "value": "stem-cell-therapy"
  },
  {
    "label": "Tumor Marker Blood Testing",
    "value": "tumor-marker-blood-testing"
  }
] as const